Use ModelMultipleChoiceFilter instead of ModelInFilter
The ModelInFilter seems to be used mostly in REST APIs, so
I have left it as-is. For GraphQL APIs, it seems better to
use ModelMultipleChoiceFilter which allows filtering by
a list of PKs instead of having to give them as a string
like "1,2,3".

This will also make the filters compatible with the latest
version of Graphene-Django (2.15.0 and above).