Refactor ReservationCreateSerializer validate method
MOve the validate method upper and have all the `check` methods below that.

Also reorders the checks per reservation unit so that the scheduler ones
are called last. This reduces the amount of mocking job needs to do in tests for
openinhours and slightly improves performance 'cause no need to contact
outside services (hauki)

TILA-1639