Add payment_type support to confirm
This allows confirmers to define the payment type.
There is a validation logic that makes sure that only
suitable values are allowed. The field is optional so
if it is missing, the code uses the payment type from
reservation unit. There is also a logic how payment
type is selected when there are multiple options
available.