Add hauki_department_id property to unit
Hauki uses the department id with the prefix of the origin_id. In this
department id case it needs to in form "tprek:<id>" to make the link
work properly.
This adds computed property called "hauki_department_id" to Unit model which returns the
tprek_department id with prefix "tprek:".

Refs TILA-2283