ReservationState is closed when begin and end past and same
ReservationUnitReservationStateHelper returned reservation unit's
reservation state as None when the reservation begi nand end was in past
and was the same.
In this scenario it should be closed state.

This changes the ReservationUnitReservationStateHelper to return
reservation state as RESERVATION CLOSED when reservation unit's
reservation begins and *_ends dates are in past and the same

Refs TILA-2451