Add a new required field
Web shop broke the API contract and added a field that is now required.
We don't have this field in our model because we don't need it. So, I just
added it with a hard-coded whitespace value. Empty string is not allowed
on web shop side. If we ever need the field, we can add it to the model
and replace the hard-coded value.

Refs TILA-2542