Use a shared cache in ghcr.io instead of github actions
Github actions cache is isolated per-branch, which means cache would
only be used on the same PR for instance. This configuration shares
cache between all runs of the workflow, which means we usually get the
lower build levels cached, but will rebuild the upper levels more often.
Still, if the image layers are designed with this in mind, the impact
of rebuilding the upper layers should be minimal.