fix: notifications form validation
Validate date and time only if not a draft or it's set.
Validate invalid dates correctly.
Add translations to validation errors.
Start time can be in the past for both drafts and non drafts
End time can be in the past for drafts.
Unset dates for drafts if they are not provided.