Add two new lookups
`__in_range` for the integer field, which takes a python range object
and compares if the value is found in that range. This is more optimal
for big ranges since the lookup converts to generate_series function
in a subquery rather than a big list of parameters.

`__seconds` for the duration field, which can be used to convert
a duration/timedelta to seconds. This can be helpful if there are
some further operations that are easier with integers.