Optimize affecting time spans fetching for FRT calc
Instead of fetching affecting time span elements with a prefetch,
create a custom QuerySet iterator method that can fetch the
time spans in chunks when a new iterator chunk needs to be fetched.
This way we also don't create so many duplicates for in memory.
The method seems much faster (up to 4x), executing in less than a
second with our current test dataset.