Add a flag for skipping `iat` validation locally
This is a hack that can be enabled for local development if you get
errors during authentication for "The token is not yet valid (iat)".
These happens due to incorrect implementation for checking issued at
time in`jwt.api_jwt.PyJWT._validate_iat` (`iat` is an int while `now`
is a float,which can be a few milliseconds off if the server
authenticated too fast).