Rewrite `HaukiResourceHashUpdater`
This was done due to problems encountered on the test server, where for some reason the process was interrupted (most likely due to timeout), which led to deletion of existing future ReservableTimeSpans, but not creation of new ones since the operation was not atomic before. Now if there are issues with any one OriginHaukiResource, the data processed before it is updated.
Also fixes a discovered bug, where new data was not updated from the Hauki API unless the hash was changed