Dont refresh AffectingTimeSpans as much & dont use for series overlap
For some reason, AffectingTimeSpans.refresh() started taking longer
than usual in production, no idea why, but calling it only during
a scheduled task should be enough to make it bearable.

However, series creation/reschedule used this as a way to check
overlapping reservations. Remove this in favor of fetching reservations
directly, as AffectingTimeSpans is likely to be behind actual
data more often.